coba type data

+++ soal1.java
@@ -1,8 +1,8 @@
 public class soal1{
     public static void main(String[] args){
-        byte angka1 = 125;
-        byte angka2 = 6;
-        byte hasil = (byte) (angka1+angka2);
+        int angka1 = 125;
+        int angka2 = 6;
+        int hasil = (int) (angka1+angka2);
 
         System.out.println("Hasilnya " + hasil);
     }
